Add Render to Deployment
This is related to https://github.com/chroma-core/chroma/pull/218
It links to the chroma-core/chroma repo, which is appropriate if the above PR is merged. If you don't want the above PR but are still interested in this button, you could point the link to https://github.com/ericmand/render-chroma which does the same thing.
I put the link above AWS because the instructions are so much shorter but you could move below if you prefer.